pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "raintux"
version = "0.1.0"
description = "Rainmeter-style desktop overlays for Linux Wayland"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "RainTux contributors" }]
dependencies = [
    "pycairo>=1.26.0",
    "PyGObject>=3.48.0",
    "psutil>=5.9.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "aiohttp>=3.9.0",
    "Pillow>=10.2.0",
    "watchdog>=4.0.0",
    "rich>=13.7.0",
    "dbus-python>=1.3.2",
    "pywayland>=0.4.15",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
lua = ["lupa>=2.0"]
audio = ["pulsectl>=23.5.2"]

[project.scripts]
raintux = "raintux.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["raintux"]

[tool.hatch.build.targets.wheel.force-include]
"raintux/data" = "raintux/data"

[tool.pytest.ini_options]
asyncio_mode = "auto"
